Extract main thread custom editors to own file
